+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRenderer.java
@@ -56,10 +56,10 @@ public final class EntityTableCellRenderer {
 						.optional(attribute)
 						.orElse(null);
 
-		FilterTableCellRenderer.Builder<Entity, Attribute<?>, T> builder =
-						FilterTableCellRenderer.<Entity, Attribute<?>, T>builder(attributeDefinition.attribute().type().valueClass())
-										.uiSettings(new EntityUISettings(queryCondition))
-										.string(attributeDefinition::string);
+		FilterTableCellRenderer.Builder<Entity, Attribute<?>, T> builder = FilterTableCellRenderer.builder()
+						.<Entity, Attribute<?>, T>columnClass(attributeDefinition.attribute().type().valueClass())
+						.uiSettings(new EntityUISettings(queryCondition))
+						.string(attributeDefinition::string);
 		if (!attributeDefinition.items().isEmpty()) {
 			// Otherwise the horizontal aligment is based on the Item value type
 			builder.horizontalAlignment(FilterTableCellRenderer.HORIZONTAL_ALIGNMENT.getOrThrow());